synapse guest module: change license
Set license to multi-license of AGPL + Element Commercial, and add
associated license files to the repository root.

This does not add a notice of "This file includes modifications" to any
files because they were only moved and relicensed, but otherwise
unmodified.

The exception is run_in_venv.sh, which is a rewrite of upstream's
run_in_venv.js and thus lacks the original copyright entirely.